Online update (WIP)
The game is not able to send and receive messages for troops spawning and buildings being created, along with starting a new settings file to use for easier access. Testing needs to be done to see if an error raised when using an external IP is due to port forwarding not being enabled, and to occasionally send a large message to resync the two clients completely in case slightly differing execution times begin to make each player's game slightly offset.